add Rlox struct
