+++ src/lib.rs
@@ -1,4 +1,5 @@
 pub mod chunk;
 pub mod compiler;
+pub mod rlox;
 pub mod scanner;
 pub mod vm;
